backend/requirements.txt
werkzeug==3.1.3
flask-cors==3.0.10
gunicorn==20.1.0
azure-core==1.32.0
azure-search-documents==11.5.2
Flask==3.1.0
openai==1.65.3
python-dotenv==1.0.1
requests==2.32.3
pyodbc==5.2.0
PyJWT==2.3.0
cryptography==36.0.1
Flask-Session==0.8.0